refactor: common runtime permissions
